feat: Post 내에 getPostData와 getComments 구현 후 PostCard 컴포넌트 import
