+++ src/components/post/PostCard/PostCard.jsx
@@ -1,29 +1,32 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './PostCard.style';
 import PostContent from '../PostContent/PostContent';
-import PostAuthor from './../PostAuthor/PostAuthor';
 import UserComponent from './../../userComponent/UserComponent';
 
-const PostCard = (data) => {
+const PostCard = ({ data }) => {
   const [authorData, setAuthorData] = useState();
   const [postData, setPostData] = useState();
 
   useEffect(() => {
-    setAuthorData(data.data.author);
-    setPostData(data.data);
+    if (data) {
+      setAuthorData(data.author);
+      setPostData(data);
+    }
   }, [data]);
 
   return (
-    <S.ProfileWrap>
-      {authorData && (
-        <UserComponent
-          image={authorData.image}
-          username={authorData.username}
-          accountname={authorData.accountname}
-        />
-      )}
-      {postData && <PostContent data={postData} />}
-    </S.ProfileWrap>
+    <S.PostCardWrap>
+      <S.ProfileWrap>
+        {authorData && (
+          <UserComponent
+            image={authorData.image}
+            username={authorData.username}
+            accountname={authorData.accountname}
+          />
+        )}
+        {postData && <PostContent data={postData} />}
+      </S.ProfileWrap>
+    </S.PostCardWrap>
   );
 };
